requirements.txt
cupy-cuda111==8.4.0
cycler==0.10.0
fastrlock==0.5
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
